1. State structure refactored - 'user' entities moved inside 'login' entity 2. Actions and reducers refactored accordingly to point 1. 3. Added util methods required to work with localStorage. 4. Added save of user key to localStorage during user login dispatch after Google spreadsheet api success response.
