+++ src/reducers/index.js
@@ -1,13 +1,11 @@
 import { combineReducers } from 'redux-immutable';
 
 import routing from './routing';
-import user from './user';
 import login from './login';
 import reviews from './reviews';
 
 export default combineReducers({
     routing,
-    user,
     login,
     reviews
 });
