--- src/selectors/user.js
@@ -1,13 +1,0 @@
-import { createSelector } from 'reselect';
-
-export const getUser = state => state.get('user');
-
-export const getDisplayName = createSelector(
-    getUser,
-    user => user.get('name', 'Anonymous')
-);
-
-export const getIsFetching = createSelector(
-    getUser,
-    user => user.get('isFetching')
-);
