--- src/utils/index.js
@@ -1,5 +1,0 @@
-import randomUniqArray from 'random-uniq-array';
-
-export const randomGreeting = randomUniqArray(
-    ['Hello', 'Nice to see you', 'Welcome', 'Hey', 'Hi', 'Cześć', 'Hola', 'What an honor to welcome thee']
-);
