+++ src/utils/randomGreeting.js
@@ -1,0 +1,5 @@
+import randomUniqArray from 'random-uniq-array';
+
+export const randomGreeting = randomUniqArray(
+    ['Hello', 'Nice to see you', 'Welcome', 'Hey', 'Hi', 'Cześć', 'Hola', 'What an honor to welcome thee']
+);
